Separates MongoDB DocumentIO integration tests for easier configuration and cleaner setup
