+++ database-impl/mongodb/src/test/java/com/findwise/hydra/mongodb/DatabaseConfigurationFactory.java
@@ -8,4 +8,5 @@ public class DatabaseConfigurationFactory {
 		config.setNamespace(namespace);
 		return config;
 	}
+
 }
